pyproject.toml
[project]
name = "kiya-002"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "websockets>=15.0.1",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "pandas>=2.3.2",
    "openpyxl>=3.1.5",
]
